services/orchestrator_service/app/main.py
from fastapi import FastAPI, Request, HTTPException
from fastapi.responses import JSONResponse
import httpx
import os
import logging
from prometheus_client import make_asgi_app
from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
from opentelemetry.instrumentation.httpx import HTTPXClientInstrumentor

app = FastAPI()
FastAPIInstrumentor().instrument_app(app)
HTTPXClientInstrumentor().instrument()

# Observability endpoints
app.mount("/metrics", make_asgi_app())

@app.get("/health")
def health():
    return {"status": "ok"}

@app.get("/model/tree")
async def get_model_tree(request: Request):
    # Fan-out to all element services (stub)
    # Merge results, preserve correlation_id
    return {"tree": []}

@app.get("/model/{element}/{id}/full-details")
async def get_full_details(element: str, id: str, request: Request):
    # Fan-out to element service, aggregate traceability, KPI, etc. (stub)
    return {"element": element, "id": id, "details": {}}
