services/orchestrator_service/requirements.txt
fastapi
uvicorn
httpx
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
pytest
pytest-asyncio
flake8
bandit
